+++ lib/middleware/middleware.h
@@ -20,7 +20,7 @@ extern StrategyType activeStrategyType;
 
 void middlewareSelectStrategy(StrategyType strategyType);
 void initMiddlewareStrategyInject(void *metricStruct, size_t metricStructSize,void (*setValueFunction)(void*,void*),void (*encodeMetricFunction)(char*,size_t,void *),void (*decodeMetricFunction)(char*,void *));
-void initMiddlewareStrategyPubSub(void (*setValueFunction)(void*,void *),void (*encodeTopicFunction)(char*,size_t,void *),void (*decodeTopicFunction)(char*,int*));
+void initMiddlewareStrategyPubSub(void (*setValueFunction)(void*,void *),void (*encodeTopicFunction)(char*,size_t,void *),void (*decodeTopicFunction)(char*,int8_t *));
 void middlewareInfluenceRouting(char* dataMessage);
 void middlewareHandleMessage(char* messageBuffer, size_t bufferSize);
 void middlewareEncodeMessage(char* messageBuffer, size_t bufferSize, int type);
